net/mvpp2: don't fail port_start on tm_init fail
Not all kernels support ETH_GSET ethtool option.
This patch is a workaround, not to fail the entire eth_dev,
if the tm_init fails due to that.

In future, all marvell kernel drivers willl support ETH_GSET or
ETHTOOL_GLINKSETTINGS , this patch should be replaced with the
ability to check both.

Change-Id: I552389a0712f90980295ce22d1e3fdb44fa7a42b
Signed-off-by: Yuval Caduri <cyuval@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/55637
Reviewed-by: Natalie Samsonov <nsamsono@marvell.com>
Tested-by: Natalie Samsonov <nsamsono@marvell.com>