crypto/mvsam: add S/G support to crypto dirver
The patch adds support for chained source mbufs given
to crypto operations. The crypto engine accepts source buffer
containing a number of segments. The destination buffer
stays the same - always one segment.
On decryption, EIP engine will look for digest at 'auth_icv_offset'
offset in SRC buffer.It must be placed in the last segment and the
offset must be set to reach digest in the last segment.
If application doesn't placed digest in source mbuf, driver try to
copy it to a last segment.

Change-Id: Id09fc362f4cb488110b5b8560552690a4ef46681
Signed-off-by: Natalie Samsonov <nsamsono@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/56686
Reviewed-by: Dmitri Epshtein <dima@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/57979